Mega Laaaaaaaaaaan

+++ src/main/java/controller/owner/DeleteControl.java
@@ -4,14 +4,20 @@
  */
 package controller.owner;
 
+import dao.DBContext;
 import dao.GearDAO;
 import java.io.IOException;
 
+import dao.OwnerDAO;
+import dao.UserDaoImpl;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
+import model.User;
+
 import java.sql.Connection;
 
 /**
@@ -24,13 +30,29 @@ public class DeleteControl extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        String id = request.getParameter("id"); //get id tu jsp
-        GearDAO gear = new GearDAO(con);
-        try {
-            gear.deleteGear(id);
-        } catch (Exception ex) {
+        HttpSession session = request.getSession();
+        OwnerDAO ownerDAO = new OwnerDAO(new DBContext());
+        User auth = (User) request.getSession().getAttribute("currentUser");
 
+        if (auth == null) {
+            session.setAttribute("message", "You are not logged in!");
+            response.sendRedirect("login.jsp");
+        } else {
+            UserDaoImpl userDAO = new UserDaoImpl();
+            boolean isDeactivated = userDAO.isAccountDeactivated(auth.getId());
+            if (isDeactivated) {
+                session.setAttribute("message", "Your account has been deactivated!");
+                response.sendRedirect("loginMessage");
+            }
+
+            String id = request.getParameter("id"); //get id tu jsp
+            GearDAO gear = new GearDAO(con);
+            try {
+                gear.deleteGear(id);
+            } catch (Exception ex) {
+
+            }
+            response.sendRedirect("viewOwner");
         }
-        response.sendRedirect("viewOwner");
     }
 }
